pyproject.toml
[project]
name = "testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.17.0",
    "fastapi>=0.124.4",
    "fastapi-users-db-tortoise>=2.0.0",
    "fastapi-users>=15.0.1",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "optional>=0.0.1",
    "tortoise-orm>=0.18.1",
    "aerich>=0.8.0",
    "asyncpg>=0.31.0",
    "google-auth>=2.43.0",
]

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./app/migrations"
src_folder = "./."
